Fix PCRE2 error constants and gracefully handle disabled JIT
When the JIT is disabled, the PCRE.jit_compile function would throw
to complain. That's not particularly nice, since we don't check before
whether the JIT is enabled. Instead, return a boolean to indicate whether
the jit was enabled (other errors still throw at jit compile time).
While I was at it, I noticed that we no longer have the ERROR_ flags
in pcre_h.jl (I'm sure we once did, because there's other parts of
this file that use them, in particular the error path in PCRE.info).
Adjust the regex to also capture those error falgs, which are specified
as e.g.:
```
```
Finally, add a test for the error path in PCRE.info to make sure it
throws an ErrorException, not an UndefVarError.