Vendor newer config.sub for emscripten support in dependencies
Upstream config.sub supports emscripten as of three hours ago [1].
However, this file is generally bundled in the tarballs of our
dependencies, so it'll be some time until they have caught up.
In the meantime, we'll have to patch in the newer config.sub to be
able to build our dependencies against emscripten. Since all
autotools-based projects can use the same emscripten, we distribute
the file itself rather than trying to maintain patch files for
whatever version of config.sub our dependencies happen to ship.

It's likely that some of the other dependencies will also need this,
but for now I've only added it to pcre/gmp/mpfr, which have been
confirmed to work when built with emscripten. We can add this
patch to other dependencies as they get enabled on the emscripten
port.

[1] http://git.savannah.gnu.org/cgit/config.git/commit/?id=b3914f8d5588d1f9ce2089cf711119653fe469fc)