gf: fix jl_isa_compileable_sig again (#30458)
The last attempts were pretty good, but still missed a lot.
But this is what you find when you actually try to test it.
This is not too important, but it can reduce compilation
performance in some case, so it is not ideal.

(cherry picked from commit 76e7421cda22d64d0c7d67ebdf4043fcb526b325)