Use XCode 8.3 for macOS on Travis (#30599)
This corresponds to macOS 10.12 Sierra. XCode 8 covers El Capitan and
Sierra, so if Travis is giving us XCode 8.x for x < 3, we're on El Cap.
Homebrew supports only three versions of macOS at a time, which means
that El Cap (10.11) is no longer support. This is likely why our Mac
builds are trying to build GCC from source; a bottle might not be
available.