solves one part of massive slowdowns using stdin/stdout #30044
a unix pipe is detected as UV_NAMED_PIPE, but handling it using uv_pipe_t
leads to massive slowdowns. It is also not the correct way to handle it according to
acording to https://nikhilm.github.io/uvbook/processes.html:
"libuv's uv_pipe_t structure is slightly confusing to Unix
programmers, because it immediately conjures up | and pipe(7). But
uv_pipe_t is not related to anonymous pipes, rather it is an IPC mechanism"
treating it as a file instead removes the slowdown completely