Code for faster factorial (and gamma)
This is kind of my first Julia PR (my other PRs have only been very trivial).

Before I had a call and three checks and taken assembly branches before looking up in the table, and now one (on 1.2 at least, more on 0.5):

```
julia> @code_native factorial_lookup(3, _fact_table64, 34)
	.text
; ┌ @ REPL[200]:2 within `factorial_lookup'
	pushq	%rax
	cmpq	$16, %rdi
	jae	L16
; │ @ REPL[200]:5 within `factorial_lookup'
; │┌ @ array.jl:728 within `getindex'
	movq	(%rsi), %rax
	movq	(%rax,%rdi,8), %rax
; │└
; │ @ REPL[200]:6 within `factorial_lookup'
	popq	%rcx
	retq
; │ @ REPL[200]:3 within `factorial_lookup'
L16:
	movabsq	$julia_factorial_lookup_helper_12541, %rax
	callq	*%rax
	popq	%rcx
	retq
	nop
; └
```

So I achieved my objective, *should* be faster, didn't time, maybe this doesn't matter much, but with cumprod table generation is also simpler and backportable to 0.5 (undef blocked that before). I didn't count bytes for generated assembly (is there away? from Julia?) and since inlined it may matter?



I did have:

factorial(n::BigInt) = factorial_lookup_helper(n, _fact_table128, 34)

originally in the PR, to bypass the inlining. It turned out to be not needed it seems and wrong to have it in anyway.

julia> factorial(BigInt(34))
295232799039604140847618609643520000000

It seems this is all backportable to 1.0 (and even to 0.5; while assembly is not as tight there, but still better) while I got this different error on 0.5 compared to 1.2 for the old outdated code:

julia> factorial(BigInt(35))
ERROR: MethodError: Cannot `convert` an object of type String to an object of type OverflowError
This may have arisen from a call to the constructor OverflowError(...),
since type constructors fall back to convert methods.
 in factorial_lookup_helper(::BigInt, ::Array{UInt128,1}, ::Int64) at ./REPL[3]:3
 in factorial(::BigInt) at ./REPL[13]:1




I was also toying with a fast-path for BigInt, seem not needed:

@noinline function factorial_lookup(n::BigInt, table, lim)
    if 0 <= n < 32
        @inbounds f = table[n+1]
        return oftype(n, f)
    else
        return factorial_lookup_helper(n, table, lim)
    end
end