make length(::String, ::Int, ::Int, ::Int) private again (#30965)
This was unintentionally made public by the great `_length` rename:

e99191a667e2a14c34186f54faed4e474943796c

Arrays and strings were using the `_length` name for private stuff
in different ways and the refactor made this part of the `length`
API when it should have remained private. Fortunately, it's not
documented and it seems unlikely that anyone is using it.

Also rename `next_continued` to `iterate_continued` to match the
name of the method that it's continuing.