Allow removing non-inlined functions that are proven pure
Right now all of our effect-free modeling that forms the basis
of our dead-code elimination operates on intrinsics (since we now
have pretty good effect free modeling for those). Of course, this
means that up until now dce has been tied to inlinability. That
isn't great, because there are lots of otherwise effect-free things
that we don't want to inline (e.g. matrix multiply is effect free
in a certain sense, and it would be nice to eliminate matmuls whose
result isn't used). This commit starts us down that track by marking
things we have proven to be pure as effect-free.