Fix a missing widenconst in SROA
We can now infer Expr(:new) as returning a `PartialStruct` if
one of the arguments is constant. However, SROA was missing a
corresponding widenconst and thus refusing to process such Expr(:new).
This didn't show up very much in julia code, because inlining generally
dropped the PartialStruct annotation on the Expr(:new), but it does
show up with more aggressive inlining schemes that I'm playing with.
Fix it and add a test.