Improve searchsorted{,first,last} examples in documentation
Use the same inputs for all three functions, and cover various cases, to help
the reader grasp how these functions work and relate to each other.

Also remove the examples using the `rev` option, which did not really make
sense, as the input was not reversed, so these really were examples on "what
happens if the input is not sorted", which does not really help much (if at
all, an explanation in the text would be better).