Properly propagate taint in SROA
When SROA crosses the compaction line to look at uncompacted IR,
it needs to be sure to propagate that taint to all values it pulls
out of the uncompacted IR. We had this logic for phi nodes, but
we accidentally lacked it for raw SSAValues. This didn't cause
problems in practice, because our pipeline compacts out raw
SSAValues before running SROA, but it is otherwise legal.
Fix that and add a test.