Prevent domtree construction from overflowing for large functions (#31531)
Switch domtree construction to a worklist

One of the first things you learn when working on a production compiler
is that you should never do recursion over the structure of the input
(because the input tends to be significantly larger). Now, I ignored
that in a couple of places when writing the optimizer, because let's
face it, Julia functions don't tend to be THAAT big (for values of THAAT
around 200k basic blocks), because we limit inlining. The TPU backend
however, reuses the julia compiler infrastructure with much more aggressive
inlining settings, making function that large not uncommon. Switch
to a worklist.

Co-Authored-By: vchuravy <vchuravy@users.noreply.github.com>