Allow a callback to update the displayed location of a method (#31514)
This is for better integration with Revise.jl.

Note: Method is mutable, so why not just update the file & line fields?
Location info gets compiled into code on a line-by-line basis;
to compute the offset into the method body, you need to know the
method's location at the time the method was compiled. Consequently we
leave those fields untouched so that the correct offsets can be computed.