loading: simplifications of code logic (#29807)
Attempt a simplification of the module-search logic.
Previously, it was mostly using a sort of 3VL (UUID found,
UUID not-found, environment not-found). However, it turns
out that it can also be handled via nullable (e.g. is the Pkg
identified or not), which also maps more easily into a dict-like
view of things (is the Pkg identified in this set or not).