splatnew: ensure performance of struct users of it (#31550)
Types which permit nargs < nfields in their constructor are required by correctness
to take a notable performance hit. Since currently nobody is using currently that
ability (NamedTuples, for example, want it to be statically disallowed),
ensure that splatnew respects that we must have nargs >= ninitialized
in the most performance-preserving way we can.