Exec julia to determine its version
When a julia executable is found, if it's not in a framework then there
is no way to know what version it is.  To determine the version, run a
simple julia expression to print the version.  Of course, running any
random executable named 'Julia' isn't smart so the executable is instead
run inside an App Sandbox with mostly restricted permissions.  The only
special permission granted to the sandbox is the read/exec '/'
permission to allow the executable to be read and executed.