During bootstrap, print julia assertion failures using fallback printing
The `@assert` macro already had support for assertions being defined using
bootstrap. However, it would throw an uninformative `UndefVarError(:Base)`
if one of those assertions actually failed during bootstrap (when Base is
unavailable). Instead, for assertions generated during bootstrap add an
extra check to see if Base is available and if not, use the fallback printing
to print the assertion message.