Expand documentation of custom random samplers.
Following discussion on Discourse, this PR expands the documentation
for defining custom random samplers. Motivation is provided for the
design, followed by examples that gradually introduce features and
explain why/when they should be used.

No text was removed, but some parts were moved to other subsections.

Docstrings added for relevant methods and structures, and included in
the docs.

The use of pre-existing types is emphasized.