Clarification when loading modules from LOAD_PATH
This proposed change is a response to (this discussion on discourse)[https://discourse.julialang.org/t/module-loading-case-sensitivity-on-windows/18523/4]. Whilst this may be a little early in the docs to be introducing the `LOAD_PATH`, I believe I would have found this clarification useful longer term. Please do say if you think this information would be best placed elsewhere.