For comparison, just always double.
This should be the fastest possible option, I think, at the expense of
larger memory use.

On my machine, compared with master, this brings the time to push! 2^30
elements one at a time onto a 2^30-size array down from 1000s to 18s.

The results of that calculation are below:

Benchmark:
```julia
for s in (2^27, 2^30,)
    vs = samerand(s)
    g["push_single_large!", s] = @benchmarkable push!(x, $(samerand())) setup=(x = copy($vs))
    g["push_multiple_large!", s] = @benchmarkable perf_push_multiple!(x, $vs) setup=(x = copy($vs))
end
```
Results on `master`:
```
  ("push_single_large!",    134217728) => Trial(768.667 ms)
  ("push_multiple_large!",  134217728) => Trial(6.169 s)
  ("push_single_large!",   1073741824) => Trial(28.277 s)
  ("push_multiple_large!", 1073741824) => Trial(1069.542 s)
```
Results after this commit:
```
  ("push_single_large!",    134217728) => Trial(766.012 ms)
  ("push_multiple_large!",  134217728) => Trial(2.211 s)
  ("push_single_large!",   1073741824) => Trial(20.854 s)
  ("push_multiple_large!", 1073741824) => Trial(18.894 s)
```

(Note though that the very large numbers fluctuate wildly, since the
process takes more virtual memory than I have physical memory, and is
swapping out to disk.)