Remove tiers of support (#32022)
The tiers of support are now maintained in one place, i.e. on the downloads page on the website. It was difficult to have an accurate list of tiers for master, and much simpler to have support tiers for the latest stable release instead. Thus, the downloads page is the right place to host this information and curate it.