More methods for Irrational
Define `zero(Irrational)` and `one(Irrational)` to return `false`
and `true` respectively. (This is breaking if someone is currently
using `true * pi` or `pi + false` as a way of forcing conversion to
`Float64`.)

Make `false` act as the additive identity for `Irrational` by defining
a method for `+`. (Not type stable)

Make `true` act as the multiplicative identity for `Irrational` by
defining a method for `*` (Not type stable)

Make `widemul` of two `Irrational`s return a `BigFloat`

Make `sin(pi)` and `tan(pi)` return exactly 0.0

Define a number of methods on `Irrational` to default to conversion to
Float64.