Add backedges for Cassette
Adds an extra field in `CodeInfo` that allows users (Cassette, et al.)
to specify dependencies (as forward edges to `MethodInstance`s) that should
be turned into backedges once the CodeInfo is passed over by inference.

The test includes a minimal implementation of the Cassette mechansim to
exercise this code path.