improved implementation of hypot(a,b) (#31922)
* Replacement for hypot(a,b)

Provides a fast and accurate implementation of hypot() that leverages
the fused multiply add where available. The approach is explained
and tested in detail in the paper:
 An Improved Algorithm for hypot(a,b) by Carlos F. Borges
The article is available online at ArXiv at the link
https://arxiv.org/abs/1904.09481

This version gives correctly rounded results more than 99% of the time. It is 10 times better than the C math library hypot function in this respect.