Have jl_realloc_aligned /try/ realloc, then fallback to malloc if not aligned
Since there isn't a system realloc_aligned on non-windows systems, we
will simply call realloc, and hope that it will either manage to grow
the existing allocation, or if it had to move it, hope that the new
allocation is correctly aligned. If not, we will manually redo the
allocation via (malloc_aligned, copy, free).

Note that this makes growing large arrays drastically faster:

```
$ julia-master -e 'b = collect(1:2^30); @time push!(b, 1);'
  5.433267 seconds (121 allocations: 327.686 MiB, 0.02% gc time)
$ ./julia -e 'b = collect(1:2^30); @time push!(b, 1);'
  0.003441 seconds (128 allocations: 327.689 MiB, 31.64% gc time)

Down to 0.003441 seconds from the previous 5.433267 seconds to double an
array of 2^30 integers. :)