Stop detaching process for precompiling modules.
This means that interrupting the precompilation (ctrl-c) or killing
Julia will kill the process that is doing the precompilation.

This is useful because currently, interrupting julia while it is
precompiling a module will leave that precompilation running in the
background. For large packages that take a long time to precompile, this
can be frustrating, especially because the user most likely interrupted
the precompilation because they _didn't_ want it to happen (e.g. maybe
they realized they forgot to change something in the code).