Test set macro specifically for desugaring testsets
This allows for a much lighter weight way of writing a big list of test
cases as a flat list, cutting down on a lot of repetition and improving
the indentation. The grouping is perhaps a little less clear but that
seems like a reasonable price to pay for improved formatting.