Switch to comparison-based F16 conversion alogrithm on x86
It turns out that due to AVX, doing F16 conversion using comparisons
(which the compiler turns into predicated SIMD computations) is 5x faster
on Intel machines with AVX2 (likely even faster with AVX512), because
the memory lookup becomes more expensive compared to the arithmetic.
The same is not true of other architectures or when the function isn't
vectorized. E.g. on PowerPC, this seems to cause ~10% slowdown.
However, given these considerations, it's not clear we actually want to
do this. All CPUs with AVX512 also have F16C, i.e. have built-in instructions
that perform this conversion in 2 cycles, so we should probably just
turn that on instead.

Benchmark scripts: https://gist.github.com/Keno/0bf23c7d0c75cfdab84580039d54997e