Split Cmd out of process.jl
This change is part of #30534 (libuv disable flag), but I think it makes
sense generally, since Cmd and Process are two orthogonal concepts, and I
think it makes sense to have one (Cmd, a general representation of a POSIX
argument list) and not the other (e.g. in cases where you want to a command
to a remote, or more trivially to have the parsing macro available even
when you can't spawn a process).