Matrix operations involving different element types
Making operations with different element types can impact a lot execution time. It would be nice in the documentation an example showing the tradeoffs a user has to decide. I am not sure where.

I am writting this request as a remainder, feel free to change my "spanglish" style or even the examples. I think though it's a common situation that a lot of users will face and it would be nice to have a tiny part of the documentation commenting this.